Bump the lower bound for `network`
I noticed `pinch` was marked as broken on nixpkgs and so looking into it I
realized `network`s version was at 3.1.1.1 which doesn't have this function
called `openSocket` that's used in `pinch`s test suite.

While this update won't be a fix for nixpkgs it makes sense updating the lower
bound to make sure `pinch` can't use a  version of `network` without
`openSocket`.