feat: Resilient, incremental parsing
This is an attempt to make pinch resilient against bad packets.

To do this, I made parsing messages incremental, meaning that the parse
can fail before all the packets of the message have been read into
memory.

I've made decoding the method name incremental, and added the option
to decode via the spec verbatim, or to only accept method names that
can be produced by the thrift IDL compiler.