feat: Add header transport mode
This change adds support for the header format, (barely) documented
[here](https://github.com/apache/thrift/blob/master/doc/specs/HeaderFormat.md).

Thrift's header transport mode is not pretty.
It completely breaks the separation of 'transport' and 'protocol'.
Existing implementations of servers support receiving messages from 'framed' and
'unframed' transport clients. This is not so much backwards compatibility, as
sideways compatibility. Does it support all protocols supported by these
transport types? In your dreams. You only get supoprt for the 'binary' and
'compact' protocols.

So the complete list of clients that a 'header' server supports, is:

| Server protocol | Client transport | Client protocol |
| --------------- | ---------------- | --------------- |
| Binary          | Framed           | Binary          |
| Compact         | Framed           | Compact         |
| Binary          | UnFramed         | Binary          |
| Compact         | UnFramed         | Compact         |
| Anything        | Header           | Same as server  |

To support these first four cases, the 'header' transport layer inspects bytes
of the message looking for patterns which uniquely identify these
transport/protocol combinations.

This is slow and bad, but it's how it's done.